+++ Source/Scene/Cesium3DTileBatchTable.js
@@ -230,10 +230,13 @@ Cesium3DTileBatchTable.prototype.applyStyle = function (style) {
   for (var i = 0; i < length; ++i) {
     var feature = content.getFeature(i);
     var color = defined(style.color)
-      ? style.color.evaluateColor(feature, scratchColor)
+      ? defaultValue(
+          style.color.evaluateColor(feature, scratchColor),
+          DEFAULT_COLOR_VALUE
+        )
       : DEFAULT_COLOR_VALUE;
     var show = defined(style.show)
-      ? style.show.evaluate(feature)
+      ? defaultValue(style.show.evaluate(feature), DEFAULT_SHOW_VALUE)
       : DEFAULT_SHOW_VALUE;
     this.setColor(i, color);
     this.setShow(i, show);