+++ Specs/Scene/SupportedImageFormatsSpec.js
@@ -4,21 +4,15 @@ describe("Scene/SupportedImageFormats", function () {
   it("constructs with options", function () {
     var supportedImageFormats = new SupportedImageFormats({
       webp: true,
-      s3tc: false,
-      pvrtc: true,
-      etc1: false,
+      basis: false,
     });
     expect(supportedImageFormats.webp).toBe(true);
-    expect(supportedImageFormats.s3tc).toBe(false);
-    expect(supportedImageFormats.pvrtc).toBe(true);
-    expect(supportedImageFormats.etc1).toBe(false);
+    expect(supportedImageFormats.basis).toBe(false);
   });
 
   it("constructs with default values", function () {
     var supportedImageFormats = new SupportedImageFormats({});
     expect(supportedImageFormats.webp).toBe(false);
-    expect(supportedImageFormats.s3tc).toBe(false);
-    expect(supportedImageFormats.pvrtc).toBe(false);
-    expect(supportedImageFormats.etc1).toBe(false);
+    expect(supportedImageFormats.basis).toBe(false);
   });
 });