+++ server.cjs
@@ -48,8 +48,7 @@
     {
       "application/json": ["czml", "json", "geojson", "topojson"],
       "application/wasm": ["wasm"],
-      "image/crn": ["crn"],
-      "image/ktx": ["ktx"],
+      "image/ktx2": ["ktx2"],
       "model/gltf+json": ["gltf"],
       "model/gltf-binary": ["bgltf", "glb"],
       "application/octet-stream": [